Validação de exclusão de batismos passados

+++ src/main/java/com/br/kyrie/services/BaptismService.java
@@ -40,6 +40,8 @@ public class BaptismService {
 	
 	@Transactional
 	public void remove (Baptism baptism) {
-		baptismRepository.deleteById(baptism.getId());
+		if(baptism.getDate().after(new Date())) {
+			baptismRepository.deleteById(baptism.getId());
+		}
 	}
 }
